rpi/requirements.txt
kivy~=2.1.0
pymysql~=1.0.2
paho-mqtt~=1.5.0
peewee~=3.15.2
pint~=0.19.2
